어빌리티 다수 / 플랫폼 / 밑점

+++ Assets/DAZB/Scripts/Ability/Effect/AbilityGlassCannon.cs
@@ -2,7 +2,7 @@ using UnityEngine;
 
 [CreateAssetMenu(menuName = "SO/Ability/GlassCannon")]
 public class AbilityGlassCannon : AbilityEffectSO {
-    public float IncreaseAttackDamage;
+    public int IncreaseAttackDamage;
     public int DecreaseHealth = 0;
     public override void ApplyEffect()
     {
